ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.15.0
yfinance==0.2.33
pydantic==2.5.2
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.2
matplotlib==3.8.2
